+++ crates/schematic/src/schema/renderers/yaml_template.rs
@@ -119,6 +119,10 @@ impl SchemaRenderer<String> for YamlTemplateRenderer {
         let mut out = vec![];
 
         for (name, field) in &structure.fields {
+            if field.flatten {
+                continue;
+            }
+
             self.ctx.push_stack(name);
 
             if self.ctx.is_hidden(field) {