+++ crates/schematic/src/schema/renderers/toml_template.rs
@@ -172,6 +172,10 @@ impl SchemaRenderer<String> for TomlTemplateRenderer {
         let mut out = vec![];
 
         for (name, field) in &structure.fields {
+            if field.flatten {
+                continue;
+            }
+
             self.ctx.push_stack(name);
 
             if !self.ctx.is_hidden(field) {